create jestSetup.js file
