+++ config/jest.config.js
@@ -27,7 +27,7 @@ export default {
   moduleNameMapper: {
     '^@shared/(.*)$': '<rootDir>/shared/$1',
     '^@ui5/webcomponents-react/dist/([Aa]ssets.*)$': '<rootDir>/packages/main/dist/$1',
-    '^@ui5/webcomponents-react/(.*)$': '<rootDir>/packages/main/src/$1',
+    '^@ui5/webcomponents-react/(.*)$': '<rootDir>/packages/main/$1',
     '^@ui5/webcomponents-react$': '<rootDir>/packages/main/src/index.ts',
     '^@ui5/webcomponents-react-base$': '<rootDir>/packages/base/src/index.ts',
     '\\.(css|less)$': 'identity-obj-proxy'
