+++ config/jestsetup.ts
@@ -1,7 +1,7 @@
 import contentLoaderSerializer from '@shared/tests/serializer/content-loader-serializer.cjs';
 import '@testing-library/jest-dom';
+import '@ui5/webcomponents-react/jestSetup.js';
 import ResizeObserver from 'resize-observer-polyfill';
-import '@ui5/webcomponents-react/dist/Assets.js';
 
 const DEFAULT_REACT_VERSION = '18';
 
@@ -50,62 +50,5 @@ jest.mock('@testing-library/react', () => {
   return jest.requireActual(packages[version]);
 });
 
-const setupMatchMedia = () => {
-  Object.defineProperty(globalThis, 'matchMedia', {
-    writable: true,
-    value: jest.fn().mockImplementation((query) => {
-      const maxWidth = parseInt(/max-width:(?<maxWidth>\d+)px/.exec(query)?.groups?.maxWidth);
-      const minWidth = parseInt(/min-width:(?<minWidth>\d+)px/.exec(query)?.groups?.minWidth);
-
-      let matches =
-        (minWidth ? minWidth <= window.innerWidth : true) && (maxWidth ? window.innerWidth <= maxWidth : true);
-
-      if (query === '(orientation: landscape)') {
-        matches = window.innerWidth > window.innerHeight;
-      }
-
-      return {
-        matches,
-        media: query,
-        onchange: null,
-        addListener: jest.fn(), // deprecated
-        removeListener: jest.fn(), // deprecated
-        addEventListener: jest.fn(),
-        removeEventListener: jest.fn(),
-        dispatchEvent: jest.fn()
-      };
-    })
-  });
-};
-
-const consoleError = console.error;
-console.error = (message: string | Error, ...args) => {
-  if (typeof message === 'string' && message.includes('Error: Could not parse CSS stylesheet')) {
-    return;
-  }
-  if ((message as Error)?.message?.includes('Could not parse CSS stylesheet')) {
-    return;
-  }
-
-  consoleError(message, ...args);
-};
-
-beforeEach(async () => {
-  (window as any).ResizeObserver = ResizeObserver;
-  window.scrollTo = jest.fn();
-  setupMatchMedia();
-  const mockIntersectionObserver = jest.fn();
-  mockIntersectionObserver.mockReturnValue({
-    observe: () => null,
-    unobserve: () => null,
-    disconnect: () => null
-  });
-  window.IntersectionObserver = mockIntersectionObserver;
-  Object.defineProperty(globalThis, 'crypto', {
-    value: {
-      randomUUID: () => `1337`
-    }
-  });
-});
-
+window.ResizeObserver = ResizeObserver;
 expect.addSnapshotSerializer(contentLoaderSerializer);
