+++ packages/cra-template-seed/template/src/setupTests.js
@@ -3,36 +3,7 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom/extend-expect';
+import '@ui5/webcomponents-react/jestSetup.js';
 import ResizeObserverPolyfill from 'resize-observer-polyfill';
 
-const setupMatchMedia = () => {
-  Object.defineProperty(window, 'matchMedia', {
-    writable: true,
-    value: jest.fn().mockImplementation((query) => {
-      const maxWidth = parseInt(/max-width:(?<maxWidth>\d+)px/.exec(query)?.groups?.maxWidth);
-      const minWidth = parseInt(/min-width:(?<minWidth>\d+)px/.exec(query)?.groups?.minWidth);
-
-      let matches = (minWidth ? minWidth <= window.innerWidth : true) && (maxWidth ? window.innerWidth <= maxWidth : true);
-
-      if (query === '(orientation: landscape)') {
-        matches = window.innerWidth > window.innerHeight;
-      }
-
-      return {
-        matches,
-        media: query,
-        onchange: null,
-        addListener: jest.fn(), // deprecated
-        removeListener: jest.fn(), // deprecated
-        addEventListener: jest.fn(),
-        removeEventListener: jest.fn(),
-        dispatchEvent: jest.fn(),
-      };
-    }),
-  });
-};
-
-beforeEach(() => {
-  setupMatchMedia();
-  window.ResizeObserver = ResizeObserverPolyfill;
-});
+window.ResizeObserver = ResizeObserverPolyfill;
